Refactors CI workflow for faster deployments
Streamlines the CI workflow to improve build and deployment times.

This commit replaces the previous method of creating a timestamped zip file containing production-ready code with a simpler approach that directly zips the necessary directories and files. This eliminates the intermediate `prod_bundle` step and reduces the time required for dependency installation.  Also, it omits the separate install dependencies step since that is handled by npm ci.